+++ src/main/java/com/project/app_dog_back/application/mapper/IUnidadMasaMapper.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.application.mapper;
 
-import com.project.app_dog_back.application.dto.UnidadMasaDto;
+import com.project.app_dog_back.application.dto.catalog.UnidadMasaDto;
 import com.project.app_dog_back.domain.model.catalog.UnidadMasa;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
