+++ src/main/java/com/project/app_dog_back/domain/service/IUnidadMasaService.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.domain.service;
 
-import com.project.app_dog_back.application.dto.UnidadMasaDto;
+import com.project.app_dog_back.application.dto.catalog.UnidadMasaDto;
 import com.project.app_dog_back.domain.model.catalog.UnidadMasa;
 
 import java.util.List;
